btrfs-progs: use unsigned size type in check_mounted_where()
Callers pass size_t and strncpy_null() also accepts size_t so change the
int parameter to match.

+++ common/open-utils.c
@@ -53,7 +53,7 @@ static int blk_file_in_dev_list(struct btrfs_fs_devices* fs_devices,
 	return 0;
 }
 
-int check_mounted_where(int fd, const char *file, char *where, int size,
+int check_mounted_where(int fd, const char *file, char *where, size_t size,
 			struct btrfs_fs_devices **fs_dev_ret, unsigned sbflags,
 			bool noscan)
 {
