+++ common/open-utils.h
@@ -23,7 +23,7 @@
 
 struct btrfs_fs_devices;
 
-int check_mounted_where(int fd, const char *file, char *where, int size,
+int check_mounted_where(int fd, const char *file, char *where, size_t size,
 			struct btrfs_fs_devices **fs_dev_ret, unsigned sbflags,
 			bool noscan);
 int check_mounted(const char* file);
